+++ CHANGES.md
@@ -3,3 +3,4 @@
 * The first release after fork of https://github.com/diffplug/spotless
 * Gradle plugin: Visualize CR and LF symbols ([#2](https://github.com/autostyle/autostyle/issues/2))
 * Gradle plugin: make autostyleCheck/Apply tasks operate only on the files that belong to the relevant projects ([#1](https://github.com/autostyle/autostyle/issues/1))
+* Diff formatter: Use simple characters when rendering diff on Windows ([#4](https://github.com/autostyle/autostyle/issues/4))